+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Changed
+- Tensorflow logging level no longer set from Python.
 ### Fixed
 - Sort VCF before annotating in `medaka_haploid_variant`.
 - Ignore errors when deleting temporary files.